requirements.txt
flask
pymongo
gitpython
lizard
requests